Console extracted to libqnapi / utils
