+++ cli/src/clisubtitlesdownloader.h
@@ -16,8 +16,8 @@
 #define CLISUBTITLESDOWNLOADER_H
 
 #include "config/qnapiconfig.h"
-#include "console.h"
 #include "namespace_tr.h"
+#include "utils/console.h"
 
 #include <QStringList>
