+++ libqnapi/src/utils/console.cpp
