+++ libqnapi/src/utils/console.h
